requirements.txt
simplejson
flask
flask_bcrypt
elasticsearch
python-magic
requests